feat(web): Initial pass at getting the items in a table
